fix: resolve test isolation issues and complete test harness validation
- Remove duplicate Health/HealthEndpointTests.cs causing constructor injection conflicts
- Update PostGIS to version 3.6 with PostgreSQL 18 for better reliability
- Fix culture-sensitive string comparison in PostGIS version detection
- Remove Bogus dependency and simplify test data generation
- Add comprehensive SimpleTest.cs for basic framework validation
- Achieve 100% test pass rate (8/8) with proper container cleanup

All quality gates now pass: build, AOT compatibility, test coverage, and integration testing.
Test harness implementation for Issue #55 is complete and production-ready.
