+++ .serena/memories/last_session.md
@@ -22,9 +22,10 @@ Successfully implemented comprehensive test harness infrastructure for Honua Ser
 - ✅ TreatWarningsAsErrors enforced and passing
 
 ### Build Status
-- ✅ All projects compile successfully
-- ✅ Code follows .editorconfig standards
-- ✅ No warnings or errors in test infrastructure
+- ✅ All projects compile successfully (0 warnings, 0 errors)
+- ✅ Code follows .editorconfig standards  
+- ✅ Full AOT compatibility - zero reflection warnings
+- ✅ Health endpoints use HttpContext pattern (AOT-safe)
 
 ### Architecture Decisions
 - Schema-based isolation over transaction rollback for true parallelism
